CARDS-2509: Reference question should support conditions based on source form metadata - Rebase ontop of latest dev to pull in other reference question changes - Code cleanup:   - Refactor null checks to remove double negatives   - Move condition utils to it's own module   - Fix test form naming issues   - Clean up constant string naming and reuse utility strings when available
