+++ modules/conditions/src/main/java/io/uhndata/cards/conditions/ConditionalUtils.java
@@ -17,7 +17,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.uhndata.cards.utils;
+package io.uhndata.cards.conditions;
 
 import java.util.ArrayList;
 import java.util.List;
