+++ modules/form-completion-status/src/main/java/io/uhndata/cards/formcompletionstatus/ConditionalSectionUtils.java
@@ -31,7 +31,7 @@ import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.uhndata.cards.utils.ConditionalUtils;
+import io.uhndata.cards.conditions.ConditionalUtils;
 
 public final class ConditionalSectionUtils
 {
